Refactor base class to be entity type specific, make getPatterns() more generic
